Refactoring based on co-change dependency: Move Method refactoring of method 'unsupportedModel': from class Exceptions to ModelMapping
